fixing command line parameters and new namespace
