+++ src/Microsoft.DotNet.Asmdiff/Csv/CsvSettings.cs
@@ -4,7 +4,7 @@
 
 using System.Text;
 
-namespace Microsoft.DotNet.Asmdiff
+namespace Microsoft.DotNet.AsmDiff.CSV
 {
     public struct CsvSettings
     {
