+++ src/Microsoft.DotNet.Asmdiff/Csv/DiffMemberCsvColumn.cs
@@ -5,7 +5,7 @@
 using Microsoft.Cci;
 using Microsoft.Cci.Mappings;
 
-namespace Microsoft.DotNet.Asmdiff
+namespace Microsoft.DotNet.AsmDiff.CSV
 {
     public sealed class DiffMemberCsvColumn : DiffCsvColumn
     {
