+++ src/Microsoft.DotNet.Asmdiff/Csv/DiffObsoletionCsvColumn.cs
@@ -7,7 +7,7 @@ using Microsoft.Cci;
 using Microsoft.Cci.Extensions;
 using Microsoft.Cci.Mappings;
 
-namespace Microsoft.DotNet.Asmdiff
+namespace Microsoft.DotNet.AsmDiff.CSV
 {
     public sealed class DiffObsoletionCsvColumn : DiffCsvColumn
     {
