+++ src/Microsoft.DotNet.Asmdiff/Csv/DiffOldAssemblyCsvColumn.cs
@@ -2,12 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.DotNet.Asmdiff
+namespace Microsoft.DotNet.AsmDiff.CSV
 {
-    public sealed class DiffNewAssemblyCsvColumn : DiffAssemblyCsvColumn
+    public sealed class DiffOldAssemblyCsvColumn : DiffAssemblyCsvColumn
     {
-        public DiffNewAssemblyCsvColumn(DiffConfiguration diffConfiguration)
-            : base(diffConfiguration, 1)
+        public DiffOldAssemblyCsvColumn(DiffConfiguration diffConfiguration)
+            : base(diffConfiguration, 0)
         {
         }
     }
