+++ src/Microsoft.DotNet.Asmdiff/Csv/DiffReturnTypeCsvColumn.cs
@@ -6,7 +6,7 @@ using Microsoft.Cci;
 using Microsoft.Cci.Extensions;
 using Microsoft.Cci.Mappings;
 
-namespace Microsoft.DotNet.Asmdiff
+namespace Microsoft.DotNet.AsmDiff.CSV
 {
     public sealed class DiffReturnTypeCsvColumn : DiffCsvColumn
     {
