+++ src/Microsoft.DotNet.Asmdiff/Csv/DiffTypeCsvColumn.cs
@@ -2,39 +2,41 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.Cci;
 using Microsoft.Cci.Extensions;
 using Microsoft.Cci.Mappings;
 
-namespace Microsoft.DotNet.Asmdiff
+namespace Microsoft.DotNet.AsmDiff.CSV
 {
-    public sealed class DiffIdCsvColumn : DiffCsvColumn
+    public sealed class DiffTypeCsvColumn : DiffCsvColumn
     {
-        public DiffIdCsvColumn(DiffConfiguration diffConfiguration)
+        public DiffTypeCsvColumn(DiffConfiguration diffConfiguration)
             : base(diffConfiguration)
         {
         }
 
         public override string Name
         {
-            get { return "ID"; }
-        }
-
-        public override string GetValue(NamespaceMapping mapping)
-        {
-            var namespaceDefinition = mapping.Representative;
-            return namespaceDefinition.DocId();
+            get { return "Type"; }
         }
 
         public override string GetValue(TypeMapping mapping)
         {
             var typeDefinition = mapping.Representative;
-            return typeDefinition.DocId();
+            return GetTypeName(typeDefinition);
         }
 
         public override string GetValue(MemberMapping mapping)
         {
             var typeDefinitionMember = mapping.Representative;
-            return typeDefinitionMember.DocId();
+            var typeDefinition = typeDefinitionMember.ContainingTypeDefinition;
+            return GetTypeName(typeDefinition);
+        }
+
+        private static string GetTypeName(ITypeDefinition typeDefinition)
+        {
+            return typeDefinition.GetTypeName(NameFormattingOptions.OmitContainingNamespace |
+                                              NameFormattingOptions.TypeParameters);
         }
     }
 }
