+++ src/Microsoft.DotNet.Asmdiff/Csv/DiffTypeIsExposedCsvColumn.cs
@@ -5,7 +5,7 @@
 using Microsoft.Cci.Extensions;
 using Microsoft.Cci.Mappings;
 
-namespace Microsoft.DotNet.Asmdiff
+namespace Microsoft.DotNet.AsmDiff.CSV
 {
     public sealed class DiffTypeIsExposedCsvColumn : DiffCsvColumn
     {
