+++ src/Microsoft.DotNet.Asmdiff/DiffConfigurationExtensions.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Cci;
 
-namespace Microsoft.DotNet.Asmdiff
+namespace Microsoft.DotNet.AsmDiff
 {
     public static class DiffConfigurationExtensions
     {
